Allows setting the imagePullPolicy for containers
The image pull policy for both apps and tasks can now be set
at deployment time by setting

    spring.cloud.deployer.kubernetes.imagePullPolicy

to one of Always,IfNotPresent,Never for each app/task.

The server can be set up to have a default image pull policy by setting

    spring.cloud.deployer.kubernetes.imagePullPolicy

to one of the given values for the server. The default value is Always.