Allows separated settings for resource limits and requests for containers
Adds the possiblity to set Kubernetes resource request values additional
to the resource limits. See [0] and [1] for an explanation

They can be set serverwide and per deployment request.

The Kubernetes resource limit values get a new key to be set but for
backwards compability the old names are still available (but have lower
priority).

I.e. with this commit the following is possible to allow burstable containers
in Kubernetes:

    spring.cloud.deployer.kubernetes.requests.cpu=500m
    spring.cloud.deployer.kubernetes.requests.memory=640Mi
    spring.cloud.deployer.kubernetes.limits.cpu=2000m
    # Same as spring.cloud.deployer.kubernetes.cpu=2000m
    spring.cloud.deployer.kubernetes.limits.memory=2048Mi
    # Same as spring.cloud.deployer.kubernetes.memory=2048Mi

[0] https://github.com/kubernetes/kubernetes/blob/release-1.3/docs/design/resource-qos.md
[1] http://kubernetes.io/docs/user-guide/compute-resources/#how-pods-with-resource-requests-are-scheduled