Fixed integration tests. Reverted commit error which removed the volume mounts being added to container.
Some KubernetesAppDeployerIntegrationTests tests were broken because the deployer properties
passed to KubernetesAppDeployer were not set on the deployer properties reference that
the container factory had when being injected into the test.
I.e. the container factory and KubernetesAppDeployer had two different instances
of the deployer properties. This was as a result of moving the handling of
volume mounts to DefaultContainerFactory. Now each test instantiates the container factory
instead of relying on injection.